requirements.txt
click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6
python-dotenv
marshmallow
flask-marshmallow
marshmallow-sqlalchemy
apispec
apispec-webframeworks
pytest==7.0.1
pyyaml